Refactor config.go position
Move to upper level
